cleaned up stuff
